requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
base58==2.1.1
beautifulsoup4==4.12.2
bidict==0.22.1
bitarray==2.8.1
bleach==6.0.0
blinker==1.6.2
boto3==1.28.35
botocore==1.31.35
Brotli==1.1.0
cachelib==0.9.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cytoolz==0.12.2
Deprecated==1.2.14
dnspython==2.3.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.1.3
Flask==2.2.5
Flask-Caching==2.0.2
Flask-Compress==1.13
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Flask-OAuthlib==0.9.6
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
fredapi==0.5.1
frontmatter==3.0.7
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
hexbytes==0.3.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==6.0.1
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
limits==3.6.0
lru-dict==1.2.0
magic-admin==1.0.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
numpy==1.21.6
oauthlib==2.1.0
openai==0.28.1
ordered-set==4.1.0
packaging==23.1
pandas==1.3.5
parsimonious==0.8.1
pkgutil_resolve_name==1.3.10
plotly==5.15.0
pluggy==1.2.0
protobuf==3.19.5
pycryptodome==3.19.0
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.5.0
pytest==7.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
python-engineio==4.7.1
python-frontmatter==1.0.0
python-socketio==5.9.0
pytz==2023.3
PyYAML==5.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.1.0
rich==13.6.0
rlp==2.0.1
rpds-py==0.10.3
s3transfer==0.6.2
scikit-learn==1.0.2
scipy==1.7.3
simple-websocket==0.10.1
six==1.16.0
soupsieve==2.5
stripe==6.6.0
tenacity==8.2.2
threadpoolctl==3.1.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
varint==1.0.2
watchdog==3.0.0
web3==5.31.4
webencodings==0.5.1
websockets==9.1
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
